USER - 카테고리, 가게정보, 장바구니, 찜

+++ src/main/java/kr/co/jhta/app/delideli/user/account/domain/UserAccount.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class UserAccount {
-    private Long userKey;
+    private int userKey;
     private String userId;
     private String userPw;
     private String userName;
